#112 Removed intrusive dialogs for toasts
- Removed flutter_easy_dialogs, positioned_dialog_manager and full_screen_dialog_manager
- Added fl_toast

- Updated showFailureDialog and showSuccessDialog to take in BuildContext and use toasts to show error or success messages
- Updated all references to the two dialog methods to provide context

- Added watch for delete and editTransferTransactionProvider
