+++ lib/presentation/debug/debug_transactions/debug_transactions_screen.dart
@@ -8,7 +8,7 @@ import 'package:pecunia/features/accounts/domain/entities/account.dart';
 import 'package:pecunia/features/accounts/usecases/get_all_accounts.dart';
 import 'package:pecunia/features/transactions/domain/entities/transaction.dart';
 import 'package:pecunia/features/transactions/usecases/create_transaction.dart';
-import 'package:pecunia/presentation/dialogs/pecunia_dialogs.dart';
+import 'package:pecunia/presentation/widgets/pecunia_dialogs.dart';
 import 'package:pecunia/presentation/widgets/transactions/forms/create_txn_form_widget.dart';
 import 'package:pecunia/presentation/widgets/transactions/recent_txn_list.dart';
 
@@ -20,12 +20,14 @@ class DebugTransactionsScreen extends ConsumerWidget {
     ref.listen(createTransactionProvider, (previous, next) {
       if (next is AsyncError) {
         ref.read(pecuniaDialogsProvider).showFailureDialog(
+              context: context,
               title: 'Oopsies',
               failure: next.error as Failure?,
             );
       }
       if (next is AsyncData<Option<Unit>> && next.value.isSome()) {
         ref.read(pecuniaDialogsProvider).showSuccessDialog(
+              context: context,
               title: 'Transaction successfully created!',
             );
       }
