+++ lib/presentation/widgets/categories/category_bottom_sheet_widget.dart
@@ -4,8 +4,8 @@ import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:pecunia/features/categories/domain/entities/category.dart';
 import 'package:pecunia/features/categories/usecases/delete_category.dart';
-import 'package:pecunia/presentation/dialogs/pecunia_dialogs.dart';
 import 'package:pecunia/presentation/widgets/categories/forms/edit_category_form_widget.dart';
+import 'package:pecunia/presentation/widgets/pecunia_dialogs.dart';
 
 void showCategoryBottomSheet(BuildContext context, Category category) {
   showModalBottomSheet<void>(
